commit on 25-04-2020 7:35pm

+++ Online/src/main/java/com/example/demo/repository/TestRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.entity.Test;
 
 @Repository
-public interface TestRepository extends JpaRepository<Test,Long>{
+public interface TestRepository extends JpaRepository<Test,Integer>{
 
 }
